+++ crates/bcr-wallet-cli/src/main.rs
@@ -1,8 +1,13 @@
-use std::str::FromStr;
+use std::{path::PathBuf, str::FromStr};
 
 use anyhow::Result;
-use bcr_wallet_core::AppState;
+use bcr_wallet_core::{
+    AppState,
+    config::{AppStateConfig, SameMintSafeMode},
+    generate_random_mnemonic,
+};
 use clap::{Parser, Subcommand};
+use nostr_sdk::RelayUrl;
 use serde::{Deserialize, Serialize};
 use tracing::info;
 use tracing_subscriber::{
@@ -15,9 +20,11 @@ mod command;
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct WalletSettings {
     pub mint_url: cashu::MintUrl,
-    pub mnemonic: String,
+    pub mnemonic: bip39::Mnemonic,
     pub log_level: String,
-    pub db_path: String,
+    pub db_path: PathBuf,
+    pub network: bitcoin::Network,
+    pub nostr_relays: Vec<RelayUrl>,
 }
 
 #[derive(Parser)]
@@ -39,6 +46,8 @@ enum Commands {
     Clear { id: usize },
     #[command(name = "add_wallet")]
     AddWallet,
+    #[command(name = "delete_wallet")]
+    DeleteWallet { id: usize },
     #[command(name = "restore_wallet")]
     RestoreWallet,
     #[command(name = "receive")]
@@ -101,7 +110,19 @@ async fn main() -> Result<()> {
 
     println!("{LOGO}");
 
-    let app_state = AppState::initialize(&settings.db_path).await?;
+    let app_state_cfg = AppStateConfig {
+        db_path: settings.db_path.clone(),
+        network: settings.network,
+        nostr_relays: settings.nostr_relays.clone(),
+        mnemonic: settings.mnemonic.clone(),
+        default_mint_url: settings.mint_url.clone(),
+        same_mint_safe_mode: SameMintSafeMode::Disabled,
+        // Disabled for now until Clowder stabilizes more
+        // same_mint_safe_mode: SameMintSafeMode::Enabled {
+        //     expiration: chrono::TimeDelta::minutes(15),
+        // },
+    };
+    let app_state = AppState::initialize(app_state_cfg).await?;
 
     match cli.command {
         Commands::Info => {
@@ -136,14 +157,21 @@ async fn main() -> Result<()> {
             info!(
                 "Adding wallet for {}: {}",
                 cli.wallet,
-                command::cmd_add_wallet(&app_state, &settings, &cli.wallet).await?
+                command::cmd_add_wallet(&app_state, &cli.wallet).await?
+            );
+        }
+        Commands::DeleteWallet { id } => {
+            info!(
+                "Deleting wallet for {}: {}",
+                cli.wallet,
+                command::cmd_delete_wallet(&app_state, &cli.wallet, id).await?
             );
         }
         Commands::RestoreWallet => {
             info!(
                 "Restoring wallet for {}: {}",
                 cli.wallet,
-                command::cmd_restore_wallet(&app_state, &settings, &cli.wallet).await?
+                command::cmd_restore_wallet(&app_state, &cli.wallet).await?
             );
         }
         Commands::RequestPayment {
@@ -194,7 +222,7 @@ async fn main() -> Result<()> {
             );
         }
         Commands::GenMnemonic => {
-            info!("{}", app_state.generate_random_mnemonic(12));
+            info!("{}", generate_random_mnemonic(12));
         }
         Commands::Recover { .. } => {
             info!("Recover for {}: NOT IMPLEMENTED", cli.wallet,);